still works -- with Federicos suggestions
